move cli to separate module
